[refactoring the code in accordance to the other tutorials]
