+++ openfl-tutorials/interactive_api/PyTorch_BrainMRI_UNet/envoy/start_envoy.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+fx envoy start -n env_one --disable-tls --envoy-config-path envoy_config.yaml -dh localhost -dp 50050
