--- openfl-tutorials/interactive_api/PyTorch_BrainMRI_UNet/envoy_1/envoy_config.yaml
@@ -1,10 +1,0 @@
-params:
-  cuda_devices: []
-
-shard_descriptor:
-  template: shard_descriptor.brainMRIShardDescriptor
-  params:
-    train_csv_path: "./train_data.csv"
-    valid_csv_path: "./valid_data.csv"
-    rank_worldsize: "1,2"
-    crop_dims: 128
