--- openfl-tutorials/interactive_api/PyTorch_BrainMRI_UNet/envoy_2/start_envoy.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -e
-
-fx envoy start -n env_two --disable-tls --envoy-config-path envoy_config.yaml -dh localhost -dp 50050
